refactor: extract shared types
