+++ src/utils/date.ts
@@ -1,4 +1,4 @@
-import type { Vacancy, Settings } from "../App";
+import type { Vacancy, Settings } from "../types";
 
 export const isoDate = (d: Date) => `${d.getFullYear()}-${String(d.getMonth()+1).padStart(2,'0')}-${String(d.getDate()).padStart(2,'0')}`;
